A guideline for errors
Errors are a crucial part of the developer experience when using an API. As
developers learn the API they inevitably run into errors. The quality and
consistency of the error messages returned to them will play a large part
in how quickly they can learn the API, how they can be more effective with
the API, and how much they enjoy using the API.

We need consistency across all services for the error format returned in
the response body.

Change-Id: Iae6dc4d84c11c95a97d2451668ebffa80c9821ec