Adding guideline for evaluating api changes
This change converts the wiki page for evaluating api changes[1] into a
guideline. The content of the page has been kept largely the same aside
from a few grammar and consistency changes. One example has been removed
as the link to the bug and review had been abandoned. The questions have
also been removed as these seemed out of place for a guideline.

[1]: https://wiki.openstack.org/wiki/APIChangeGuidelines

Change-Id: I5485970686e352d5146bcfe2a3cf53388b929b04