Clarify the return code when server have hard-code length limit
Clarify the case: when the request with valid format, but the request
exceeds the limited length of collection type property.

The instance of this case is tags. In nova there is limit of the
length of tags in server code.

Related patch: Ic1b2ecf9459798b14db12d449e003648c4228550

Change-Id: I7086fb8a06c47e85719044379ae3c88d797df6a4