http guideline expansion: background
This provides some additional background on the HTTP specification
inside of OpenStack. It lays the groundwork for understanding how we
map a high level application into HTTP, and some of the constraints in
doing so. The purpose is to provide a common background around what
the spec was originally targeted at and what we are using it for in
RESTful APIs so that we can be on a common ground when debating
interpretations of the least worst way to go about a thing.

Change-Id: I20e50533878c1c3dae32f72736db35da96ecbd28