add subsection around caching behavior and http
One of the reasons things like the 501 debate go around in circles is
because the caching implications aren't being thought about. Caching
and Proxies are an extremely fundamental concept in HTTP, but often
skipped over by folks new to it. It's important to call it out
especially as the CORS support and javascript in browser native
clients are going to have dramatically different behavior here than
the python requests library that most people are familiar with.

Change-Id: Ibc6eb07cb53e4dc5a8cee875d0eb356333465b1b