Change template to match our historic usage
The template had the assumption that each guideline would get its own
file. The downside of this is that it would make our table of contents
hard to navigate. I propose that each file should be a collection of
related guidelines and each guideline should be a subheader within the
file.

It's still easy to link to specific guidelines with anchor links to
subheaders.

Change-Id: I6ac1587fe28cbc815dffb8d0e1dcc14f006107fa