Added tags restrictions to the tagging guidelines
One problem with tag names was found during tag
implementation in Nova Ibc44228aeae94c17353af7fccfcfb2c11b2e9190
This problems was connected with URLs. We can't
use '/' in names because in this case we will get a incorrect
URL for tag 'abc/cba': /servers/<server_id>/tags/abc/cba
Even if '/' will be encoded like %2F. Nova will decide
that it is '/'.

But this is not a problem of Nova. Unfortunately some servers
works incorrectly with encoded character '/'. It's a problem of poor
implementation of some servers. To escape problems with handling URLs
character '/' should be forbidden in tag names.

Change-Id: I9832270b3d8210bff9cd2d9732b4d19f757f865a