Reword the note following microversion headers
I had to read the original note four times before it made sense to me.
The problem, I think, was that it talked about 'old' and 'new' - but I
don't have enough context to immediately remember which is old and which
is new. Also, the new thing was first, and my brain wanted to make them
be listed in the opposite order. It was also unclear to me for a bit
which was the one I was supposed to use.

Address those concerns by rewording the note to be clear that the first
option is the desired one, and to refer to each of them as first and
second rather than old and new.

Change-Id: I1d51572eb57278e6b21224ca60516c528de42ee1