Microversions: add next_min_version field in version body
This adds a "next_min_version" field in the versions body returned by
the root endpoint for a service using microversions. This gives a clear
signal to clients that the minimum supported microversion is going to
change. It also adds a "not_before" field that guarantees the that the
current minimum version will be supported up to that date, giving
consumers a sense of how quickly they need to update their programs.

Co-Authored-By: Ed Leafe <ed@leafe.com>

Change-Id: Ib717221ce6d162e3efca822a5a6b0a7ac689d7b0