Expand note about rfc5988 link header
An alternate that is suggested for people who cannot add link sections
to the response body is the RFC 5988 Link header.

Move it to its own section for ease of linking, and change the text from
'can consider' to 'should'.

Also, add a note about API breakage and versioning related to adding
Link headers.

Change-Id: Ide4124513e3328084c53a2122384a4e25ea6c863