Fix the checking of directories for extensions
The test for file extensions was checking everything returned by
glob.glob("guidelines/*"). The recent merge of the refactored HTTP
guidelines added a new guidelines/http/ directory, and the test was
failing because that entry didn't end with either .rst or .json. This
patch corrects that by skipping directories from that name check.

Change-Id: I0a84160015e9c1d96cb456b07c5e45fc6645df08